// core/admin-dashboard/store/post/action-types.ts
export enum ActionTypes {
  GET_POSTS = "GET_POSTS",
  GET_POST = "GET_POST",
  CREATE_POST = "CREATE_POST",
  UPDATE_POST = "UPDATE_POST",
  DELETE_POST = "DELETE_POST",
  HARD_DELETE_POST = "HARD_DELETE_POST",
  BLOCK_POST_COMMENT = "BLOCK_POST_COMMENT",
  UNBLOCK_POST_COMMENT = "UNBLOCK_POST_COMMENT",
  RESTORE_POST = "RESTORE_POST",
  GET_POST_ANALYTICS = "GET_POST_ANALYTICS",
  GET_MOST_POPULAR_POSTS_ANALYTICS = "GET_MOST_POPULAR_POSTS_ANALYTICS",
}