// core/admin-dashboard/store/post/mutation-types.ts
export enum MutationTypes {
  SET_POST = "SET_POST",
  SET_POSTS = "SET_POSTS",
  SET_POST_ANALYS_DATA = "SET_POST_ANALYS_DATA",
  UPDATE_POST_DATA = "UPDATE_POST_DATA",
  SET_MOST_POPULAR_POSTS_ANALYS_DATA = "SET_MOST_POPULAR_POSTS_ANALYS_DATA",
  SET_POST_PAGINATION = "SET_POST_PAGINATION",
  UPDATE_POST_PAGINATION = "UPDATE_POST_PAGINATION",
}